refactor: consolidate file generation logic into a reusable addFiles function
