+++ packages/nx-python/src/generators/poetry-project/generator.ts
@@ -1,9 +1,6 @@
 import {
   addProjectConfiguration,
   formatFiles,
-  generateFiles,
-  names,
-  offsetFromRoot,
   ProjectConfiguration,
   readProjectConfiguration,
   Tree,
@@ -22,6 +19,7 @@ import {
   BasePythonProjectGeneratorSchema,
 } from '../types';
 import {
+  addFiles,
   normalizeOptions as baseNormalizeOptions,
   getPyprojectTomlByProjectName,
 } from '../utils';
@@ -65,51 +63,6 @@ function normalizeOptions(
     devDependenciesProjectPkgName,
     individualPackage: !tree.exists('pyproject.toml'),
   };
-}
-
-function addFiles(tree: Tree, options: NormalizedSchema) {
-  const templateOptions = {
-    ...options,
-    ...names(options.name),
-    offsetFromRoot: offsetFromRoot(options.projectRoot),
-    template: '',
-    dot: '.',
-    versionMap: DEV_DEPENDENCIES_VERSION_MAP,
-  };
-  if (options.templateDir) {
-    generateFiles(
-      tree,
-      path.join(options.templateDir),
-      options.projectRoot,
-      templateOptions,
-    );
-    return;
-  }
-
-  generateFiles(
-    tree,
-    path.join(__dirname, 'files', 'base'),
-    options.projectRoot,
-    templateOptions,
-  );
-
-  if (options.unitTestRunner === 'pytest') {
-    generateFiles(
-      tree,
-      path.join(__dirname, 'files', 'pytest'),
-      options.projectRoot,
-      templateOptions,
-    );
-  }
-
-  if (options.linter === 'flake8') {
-    generateFiles(
-      tree,
-      path.join(__dirname, 'files', 'flake8'),
-      options.projectRoot,
-      templateOptions,
-    );
-  }
 }
 
 function updateRootPyprojectToml(
@@ -388,7 +341,7 @@ export default async function (
     projectConfiguration,
   );
 
-  addFiles(tree, normalizedOptions);
+  addFiles(tree, normalizedOptions, __dirname);
   updateDevDependenciesProject(tree, normalizedOptions);
   updateRootPyprojectToml(tree, normalizedOptions);
   await formatFiles(tree);
